+++ src/modules/handlers/screen/index.ts
@@ -1,6 +1,6 @@
 export { default as CurrentlyPlayingTrackHandler } from './curently-playing-track-handler';
 export { default as CenturionScreenHandler } from './centurion-screen-handler';
-export { default as GewisPosterScreenHandler } from './poster/gewis/gewis-poster-screen-handler';
+export { default as CarouselPosterHandler } from './poster/carousel-poster-handler';
 export { default as HubblePosterScreenHandler } from './poster/hubble/hubble-poster-screen-handler';
 export { default as HubbleClosedPosterScreenHandler } from './poster/hubble/hubble-closed-poster-screen-handler';
 export { default as HubbleLastcallPosterScreenHandler } from './poster/hubble/hubble-lastcall-poster-screen-handler';
