+++ src/modules/handlers/screen/screen-handler-settings.ts
@@ -1,11 +1,9 @@
 export interface ScreenHandlerSettings {
   RoomResponsibleLegacyScreenURL: string;
   HubblePosterScreenHandler: boolean;
-  GewisPosterScreenHandler: boolean;
 }
 
 export const ScreenHandlerSettingsDefaults: ScreenHandlerSettings = {
   RoomResponsibleLegacyScreenURL: '',
   HubblePosterScreenHandler: false,
-  GewisPosterScreenHandler: false,
 };
