+++ src/seed/seed.ts
@@ -13,7 +13,7 @@ import { LightsScene, LightsSceneEffect } from '../modules/lights/entities/scene
 import { SearchLightCreateParams } from '../modules/lights/effects/movement/search-light';
 import LightsWheelColorChannelValue from '../modules/lights/entities/lights-wheel-color-channel-value';
 import LightsWheelGoboChannelValue from '../modules/lights/entities/lights-wheel-gobo-channel-value';
-import GewisPosterScreenHandler from '../modules/handlers/screen/poster/gewis/gewis-poster-screen-handler';
+import CarouselPosterHandler from '../modules/handlers/screen/poster/carousel-poster-handler';
 import SimpleAudioHandler from '../modules/handlers/audio/simple-audio-handler';
 import LightsWheelRotateChannelValue from '../modules/lights/entities/lights-wheel-rotate-channel-value';
 import { FixedPositionCreateParams } from '../modules/lights/effects/movement/fixed-position';
@@ -46,15 +46,15 @@ export default async function seedDatabase() {
   const rootScreenService = new RootScreenService();
   await rootScreenService.createScreen({
     name: 'PCGEWISB-links',
-    defaultHandler: GewisPosterScreenHandler.name,
+    defaultHandler: CarouselPosterHandler.name,
   });
   await rootScreenService.createScreen({
     name: 'PCGEWISB-rechts',
-    defaultHandler: GewisPosterScreenHandler.name,
+    defaultHandler: CarouselPosterHandler.name,
   });
   await rootScreenService.createScreen({
     name: 'PCGEWISINFO',
-    defaultHandler: GewisPosterScreenHandler.name,
+    defaultHandler: CarouselPosterHandler.name,
   });
   await new AuthService().createIntegrationUser({ name: 'BPM-script' });
